requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pandas==2.2.2
Pillow==10.4.0
